Move chart render fn to another file
